Added namespace cascade configuration
The `cascade` property in the config object determines the order of
precedence for the namespace cascade. If the resolver fails to lookup a
module at the app level, it walks down this array, seeing if any of
these namespaces include the package.

Of course, if the resolver finds a matching module, that module can always
directly import modules from farther down the cascade if necessary.

I added as config here to provide a single-source-of-truth list, since
this information is needed in a few different spots.