Tweaked initializer loading to use cascade lookup
The `main.js` bootstrapping script seems to hack around the resolver by
manually checking the internal registry of the module loader for any
initializer modules. This was the only place I could see that didn't
use the Resolver for lookup, so the only other place that needed to be
tweaked to accomodate the cascade logic.

Addon initializers are a good place to do things like patch core files
(which the Resolver won't automatically pick up, i.e. patching
Ember.Route for default authentication redirects on all routes) and for
setting up routes in the Router.